requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
chromadb==0.4.22
sentence-transformers==2.3.1
pydantic==2.5.3
xxhash==3.4.1
numpy==1.26.3
aiofiles==23.2.1
python-dotenv==1.0.1
